requirements.txt
# ── Core NLP / ML ────────────────────────────────────────────────────────────
transformers==4.40.0
torch==2.3.0
datasets==2.19.0
scikit-learn==1.4.2
nltk==3.8.1

# ── Coleta RSS ────────────────────────────────────────────────────────────────
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2

# ── Banco de dados ────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1

# ── Orquestração ─────────────────────────────────────────────────────────────
prefect==2.19.0

# ── API ───────────────────────────────────────────────────────────────────────
flask==3.0.3
flask-cors==4.0.1

# ── Utilitários ───────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1
loguru==0.7.2
tqdm==4.66.4

# ── Rastreamento de experimentos ─────────────────────────────────────────────
mlflow==2.13.0

# ── Visualização ─────────────────────────────────────────────────────────────
matplotlib==3.8.4
pandas==2.2.2

# ── Testes ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-cov==5.0.0
